// src/types/restaurant.d.ts
export type Prices = {
  onsite: number;
  delivery?: number;
  pickup?: number;
};

export type OptionGroup = {
  id: string;
  name: string;
  optionIds: string[];
  minChoices?: number;
  maxChoices?: number;
  isTypeMulti: boolean;
  isRequired?: boolean;
};

export type Product = {
  id: string;
  name: string;
  description?: string[];
  imageUrl?: string;
  variants: Variant[];
  allergenIds?: string[];
};

export type Variant = {
  id: string;
  name?: string;
  optionGroupIds?: string[];
  prices: Prices;
  isSoldOut?: boolean;
  isVegetarian?: boolean;
  isVegan?: boolean;
  alcoholVolume?: number;
  caffeineAmount?: number;
  discountId?: string;
};

export type Allergen = {
  id: string;
  description: string;
};

export type Discount = {
  id: string;
  name: string;
  description: string;
  daysOfWeek: number[];
  from: number;
  until: number;
  discount?: number;
  discountedOptionGroupId?: string[];
};